pyproject.toml
[tool.poetry]
name = "icenetetl"
version = "0.1.0"
description = "Infrastructure for storing IceNet predictions and importing them into a database"
authors = ["James Robinson <jrobinson@turing.ac.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
azure-identity = "^1.6.0"
azure-mgmt-resource = "^19.0.0"
azure-mgmt-subscription = "^1.0.0"
azure-mgmt-storage = "^18.0.0"
coloredlogs = "^15.0.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
